requirements.txt

### `requirements.txt`
```text
python-dateutil>=2.8.2
typing-extensions>=4.0.0; python_version < '3.10'